pyproject.toml
[project]
name = "payments-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.1",
    "django>=5.2",
    "django-unfold>=0.55.3",
    "mypy>=1.15.0",
    "ruff>=0.11.6",
]
